build: moved components back to root
