+++ app/(payload)/admin/importMap.js
@@ -1,9 +1,9 @@
 import { WorkflowView as WorkflowView_f4d5503956e2d624777a2e6c0bc05047 } from 'payload-kanban-board/dist/exports/client.js'
-import { default as default_fc00ce5faae3a00a46a3d2f6f61f9574 } from '@/app/components/icon'
-import { default as default_f8f7bee941f715e8e840a33a30de0c33 } from '@/app/components/logo'
+import { default as default_f8022cf35b3d492829ec1a405bd134e9 } from '@/components/icon'
+import { default as default_7a89b5675912f3f4f82dd41bbb03d52e } from '@/components/logo'
 
 export const importMap = {
   "payload-kanban-board/dist/exports/client.js#WorkflowView": WorkflowView_f4d5503956e2d624777a2e6c0bc05047,
-  "@/app/components/icon#default": default_fc00ce5faae3a00a46a3d2f6f61f9574,
-  "@/app/components/logo#default": default_f8f7bee941f715e8e840a33a30de0c33
+  "@/components/icon#default": default_f8022cf35b3d492829ec1a405bd134e9,
+  "@/components/logo#default": default_7a89b5675912f3f4f82dd41bbb03d52e
 }
