+++ mdx-components.ts
@@ -1,6 +1,6 @@
 import { useMDXComponents as getThemeComponents } from 'nextra-theme-docs' // nextra-theme-blog or your custom theme
 import { MDXComponents } from 'nextra/mdx-components'
-import Hero from './app/components/Hero'
+import Hero from './components/Hero'
 
 // Get the default MDX components
 const themeComponents = getThemeComponents()
